requirements.txt
Django==1.11
-e git+https://github.com/Retzudo/django-emailuser.git@96f715aa02740441689024dcbcb6659101286036#egg=django_emailuser
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
pkg-resources==0.0.0
PyJWT==1.4.2
pytz==2017.2
